address cubic

+++ packages/trpc/server/routers/viewer/teams/managedEvents/managedEventReassign.schema.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 export const ZManagedEventReassignInputSchema = z.object({
-  bookingId: z.number(),
+  bookingId: z.number().int(),
 });
 
 export type TManagedEventReassignInputSchema = z.infer<typeof ZManagedEventReassignInputSchema>;
